Moved functions from function files into appropriate scripts
