+++ astronomy-pipeline/Dockerfile
@@ -7,9 +7,7 @@ RUN pip install -r requirements.txt
 
 COPY astronomy_load.py .
 COPY astronomy_extract.py .
-COPY astronomy_extract_functions.py .
 COPY astronomy_transform.py .
-COPY astronomy_transform_functions.py .
 COPY astronomy_pipeline.py .
 
 # Set the CMD to point to the Lambda handler
